rename helper to helpers
And move all the normalization functions to  "not_sure_useful"
